Integra downloads via Cobalt e restaura busca no YouTube

+++ src/commands/member/downloads/facebook.js
@@ -30,15 +30,17 @@ export default {
       throw new WarningError("O link não é do Facebook!");
     }
 
+    let data;
     try {
       await sendReply("⬇️ Baixando...");
-      const data = await downloadByCommand("facebook", fullArgs);
+      data = await downloadByCommand("facebook", fullArgs);
       await sendSuccessReact();
       await sendVideoFromFile(data.filePath);
-      await fs.rm(data.filePath, { force: true });
     } catch (error) {
       errorLog(JSON.stringify(error, null, 2));
       await sendErrorReply(error.message);
+    } finally {
+      if (data?.filePath) await fs.rm(data.filePath, { force: true });
     }
   },
 };
